requirements.txt
pyspark
apache-airflow
pandas